Renamed commands to command
Not a huge fan of pluralized directories for subclasses.
This is in preparation for molecule 2.0 work.
