+++ molecule/cli.py
@@ -42,7 +42,7 @@ Options:
 import docopt
 
 import molecule
-from molecule import commands
+from molecule import command
 from molecule import utilities
 
 
@@ -55,7 +55,7 @@ class CLI(object):
         command_args = {} if args.get('<args>') is None else args.pop('<args>')
 
         try:
-            command_module = getattr(commands, command_name)
+            command_module = getattr(command, command_name)
             command_clazz = getattr(command_module, command_name.capitalize())
         except AttributeError:
             raise docopt.DocoptExit()
