+++ molecule/command/init.py
@@ -26,7 +26,7 @@ import jinja2
 import sh
 
 from molecule import utilities
-from molecule.commands import base
+from molecule.command import base
 
 LOG = utilities.get_logger(__name__)
