+++ molecule/command/login.py
@@ -24,7 +24,7 @@ import signal
 import subprocess
 
 from molecule import utilities
-from molecule.commands import base
+from molecule.command import base
 
 LOG = utilities.get_logger(__name__)
